requirements.txt
# Core
streamlit>=1.36.0
numpy>=1.26.0
Pillow>=10.0.0
matplotlib>=3.8.0

# Computer Vision
opencv-python>=4.9.0
scikit-image>=0.22.0

# Deep Learning (CPU by default; swap torch index URL for CUDA build if needed)
torch>=2.2.0
torchvision>=0.17.0

# HuggingFace ecosystem
transformers>=4.40.0
datasets>=2.19.0
timm>=1.0.0
huggingface-hub>=0.22.0

# Object detection / YOLO
ultralytics>=8.2.0

# Utilities
scipy>=1.13.0
tqdm>=4.66.0
